Completed EconManager
Now just cleaning up old code and getting everything ready for the switch to LuckPerms
