+++ src/main/java/me/squid/eoncurrency/commands/PayCommand.java
@@ -1,7 +1,7 @@
 package me.squid.eoncurrency.commands;
 
 import me.squid.eoncurrency.Eoncurrency;
-import me.squid.eoncurrency.managers.VaultEconManager;
+import me.squid.eoncurrency.managers.EconManager;
 import me.squid.eoncurrency.utils.Utils;
 import net.milkbowl.vault.economy.Economy;
 import org.bukkit.Bukkit;
@@ -13,11 +13,11 @@ import org.bukkit.entity.Player;
 public class PayCommand implements CommandExecutor {
 
     Eoncurrency plugin;
-    VaultEconManager vaultEconManager = new VaultEconManager();
-    Economy eco = vaultEconManager;
+    EconManager econManager;
 
-    public PayCommand(Eoncurrency plugin) {
+    public PayCommand(Eoncurrency plugin, EconManager econManager) {
         this.plugin = plugin;
+        this.econManager = econManager;
         plugin.getCommand("pay").setExecutor(this);
     }
 
@@ -28,11 +28,11 @@ public class PayCommand implements CommandExecutor {
             if (args.length == 2){
                 Player target = Bukkit.getPlayer(args[0]);
                 double amount = Double.parseDouble(args[1]);
-                amount = Double.parseDouble(eco.format(amount));
+                amount = Double.parseDouble(econManager.format(amount));
 
-                if (target != null && eco.has(p, amount) && amount > 0) {
-                    eco.withdrawPlayer(p, amount);
-                    eco.depositPlayer(target, amount);
+                if (target != null && econManager.has(p, amount) && amount > 0) {
+                    econManager.withdrawPlayer(p, amount);
+                    econManager.depositPlayer(target, amount);
                     target.sendMessage(Utils.chat("&7[&b&lEonEco&r&7] &b" + p.getName() + " has sent you $" + amount));
                     p.sendMessage(Utils.chat("&7[&b&lEonEco&r&7] &bYou have sent " + target.getName() + " $" + amount));
                 } else {
