+++ src/main/java/me/squid/eoncurrency/commands/subcommands/currency/EcoGiveCommand.java
@@ -1,7 +1,7 @@
 package me.squid.eoncurrency.commands.subcommands.currency;
 
 import me.squid.eoncurrency.commands.SubCommand;
-import me.squid.eoncurrency.managers.VaultEconManager;
+import me.squid.eoncurrency.managers.EconManager;
 import me.squid.eoncurrency.utils.Utils;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
@@ -10,7 +10,7 @@ import org.bukkit.entity.Player;
 
 public class EcoGiveCommand extends SubCommand {
 
-    VaultEconManager vaultEconManager = new VaultEconManager();
+    EconManager econManager = new EconManager();
 
     @Override
     public String getName() {
@@ -34,7 +34,7 @@ public class EcoGiveCommand extends SubCommand {
                 Player target = Bukkit.getPlayer(args[1]);
                 if (target != null){
                     double amount = Double.parseDouble(args[2]);
-                    vaultEconManager.depositPlayer(target, amount);
+                    econManager.depositPlayer(target, amount);
                     sender.sendMessage(Utils.chat("&7[&b&lEonEco&r&7] &a" + target.getName() + " has received $" + amount));
                 } else {
                     sender.sendMessage(Utils.chat("&7[&b&lEonEco&r&7] &aPlayer is offline"));
